treewide: Fix compiler warnings
src/erp.c:134:10: error: comparison of integer expressions of different
signedness: 'unsigned int' and 'int' [-Werror=sign-compare]

src/eap-ttls.c:378:10: error: comparison of integer expressions of different signedness: 'uint32_t' {aka 'unsigned int'} and 'int' [-Werror=sign-compare]

+++ src/erp.c
@@ -114,8 +114,8 @@ static bool erp_tlv_iter_next(struct erp_tlv_iter *iter)
 {
 	const unsigned char *tlv = iter->tlv + iter->pos;
 	const unsigned char *end = iter->tlv + iter->max;
-	unsigned int tag;
-	unsigned int len;
+	uint8_t tag;
+	uint8_t len;
 
 	if (end - tlv < 2)
 		return false;
